feature(STOW-SR) Refactor
