+++ Packages/ohif-measurements/client/components/measurementLightTable/measurementLightTableView/measurementLightTableView.html
@@ -12,16 +12,16 @@
     {{#if hasAnyMeasurement}}
     <div class="report-area">
         {{#if saveEnabled}}
-            {{#unless noUnsavedChanges}}
+            {{#if hasUnsavedChanges}}
             <div class="unsaved-changes-alert">
                 You have unsaved changes
             </div>
-            {{/unless}}
+            {{/if}}
         {{/if}}
         {{#form (extend api=instance.api)}}
             {{#button class='btn p-x-2' action='exportCSV'}}Export CSV{{/button}}
             {{#if saveEnabled}}
-                {{#button class='btn p-x-2' action='save' disabled=(noUnsavedChanges)}}Save to SR{{/button}}
+                {{#button class='btn p-x-2' action='save' disabled=(not hasUnsavedChanges)}}Save to Archive{{/button}}
             {{/if}}
         {{/form}}
     </div>
