+++ Packages/ohif-servers/both/lib/index.js
@@ -1,1 +1,2 @@
 import './getCurrentServer.js';
+import './getWADOProxyUrl.js';
